Allow for curl in case git isn't installed
Git isn't on RHEL 5 by default, unless you have the epel repository
installed. Added in a test for git, if not, curl should be installed by
default on all the supported distributions.

Also, if you use [[ ]] you don't have to worry about quoting ${LSDIR},
but all the other ones should be quoted in case of spaces.